last ersion

+++ lib/DAO/service.dart
@@ -1,6 +1,7 @@
 import 'package:my_service/DAO/generic.dart';
 import 'package:my_service/models/service.dart';
 import 'package:my_service/utils/snack_msg.dart';
+import 'package:supabase_flutter/supabase_flutter.dart';
 
 class ServiceDAO {
   final GenericDAO<Service> _serviceDAO = GenericDAO<Service>("service");
@@ -74,15 +75,50 @@ class ServiceDAO {
     }
   }
 
-  // Delete a service
   Future<bool> deleteServiceById(String id) async {
     try {
+      // Fetch reservations related to the service
+      final reservations = await Supabase.instance.client
+          .from('reservation')
+          .select('id')
+          .eq('service_id', id);
+
+      for (var reservation in reservations) {
+        final reservationId = reservation['id'];
+
+        // Delete related 'avis' entries first
+        final avisDeleteResponse = await Supabase.instance.client
+            .from('avis')
+            .delete()
+            .eq('reservation_id', reservationId);
+
+        if (avisDeleteResponse != null) {
+          showMessage(
+              "Failed to delete related avis for reservation $reservationId",
+              isError: true);
+          return false;
+        }
+      }
+
+      // After deleting 'avis' entries, delete 'reservation' entries
+      final reservationDeleteResponse = await Supabase.instance.client
+          .from('reservation')
+          .delete()
+          .eq('service_id', id);
+
+      if (reservationDeleteResponse != null) {
+        showMessage("Failed to delete related reservations", isError: true);
+        return false;
+      }
+
+      // Proceed with deleting the service
       await _serviceDAO.delete("id", id);
       showMessage("Service supprimé avec succès");
       return true;
     } catch (e) {
       showMessage("Une erreur est survenue lors de la suppression du service",
           isError: true);
+      print(e);
       return false;
     }
   }
